move uploadFile to probe model
