+++ src/app/Models/Probe/UploadFile.ts
@@ -1,4 +1,4 @@
-import { AiurProtocal } from './AiurProtocal';
+import { AiurProtocal } from '../AiurProtocal';
 
 export class UploadFile extends AiurProtocal {
     public filePath: string;
