+++ src/app/Services/FilesApiService.ts
@@ -2,7 +2,7 @@ import { ApiService } from './ApiService';
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpEvent, HttpEventType, HttpRequest } from '@angular/common/http';
 import { Observable } from 'rxjs/';
-import { UploadFile } from '../Models/UploadFile';
+import { UploadFile } from '../Models/Probe/UploadFile';
 import { catchError, map } from 'rxjs/operators';
 import { AiurValue } from '../Models/AiurValue';
